Fixed tests

+++ src/main/kotlin/com/broughty/epl/db/repository/GameRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface GameRepository : CrudRepository<Game, Long> {
+interface GameRepository : CrudRepository<Game, Int> {
 
 }
